+++ src/bin_smart_rollup_node/main_smart_rollup_node.ml
@@ -384,8 +384,8 @@ let dump_durable_storage =
       | Error errs -> cctxt#error "%a" pp_print_trace errs)
 
 let export_snapshot
-    (data_dir, dest, no_checks, compress_on_the_fly, uncompressed) filename
-    (cctxt : Client_context.full) =
+    (data_dir, dest, no_checks, compress_on_the_fly, uncompressed, compact)
+    filename (cctxt : Client_context.full) =
   let open Lwt_result_syntax in
   let*! compression =
     match (compress_on_the_fly, uncompressed) with
@@ -396,7 +396,9 @@ let export_snapshot
     | false, true -> Lwt.return Snapshots.No
   in
   let* snapshot_file =
-    Snapshots.export ~no_checks ~compression ~data_dir ~dest ~filename
+    if compact then
+      Snapshots.export_compact ~compression ~data_dir ~dest ~filename
+    else Snapshots.export ~no_checks ~compression ~data_dir ~dest ~filename
   in
   let*! () = cctxt#message "Snapshot exported to %s@." snapshot_file in
   return_unit
@@ -406,12 +408,13 @@ let export_snapshot_auto_name =
   command
     ~group
     ~desc:"Export a snapshot of the rollup node state."
-    (args5
+    (args6
        data_dir_arg
        Cli.snapshot_dir_arg
        Cli.no_checks_arg
        Cli.compress_on_the_fly_arg
-       Cli.uncompressed)
+       Cli.uncompressed
+       Cli.compact)
     (prefixes ["snapshot"; "export"] @@ stop)
     (fun params cctxt -> export_snapshot params None cctxt)
 
@@ -420,15 +423,18 @@ let export_snapshot_named =
   command
     ~group
     ~desc:"Export a snapshot of the rollup node state to a given file."
-    (args4
+    (args5
        data_dir_arg
        Cli.no_checks_arg
        Cli.compress_on_the_fly_arg
-       Cli.uncompressed)
+       Cli.uncompressed
+       Cli.compact)
     (prefixes ["snapshot"; "export"] @@ Cli.snapshot_file_param @@ stop)
-    (fun (data_dir, no_checks, compress_on_the_fly, uncompressed) filename cctxt ->
+    (fun (data_dir, no_checks, compress_on_the_fly, uncompressed, compact)
+         filename
+         cctxt ->
       export_snapshot
-        (data_dir, None, no_checks, compress_on_the_fly, uncompressed)
+        (data_dir, None, no_checks, compress_on_the_fly, uncompressed, compact)
         (Some filename)
         cctxt)