+++ src/lib_smart_rollup_node/layer1.mli
@@ -77,6 +77,16 @@ val start :
   #Client_context.full ->
   t tzresult Lwt.t
 
+(** Same as {!start} but does not start monitoring L1 blocks. *)
+val create :
+  name:string ->
+  reconnection_delay:float ->
+  l1_blocks_cache_size:int ->
+  ?protocols:Protocol_hash.t trace ->
+  ?prefetch_blocks:int ->
+  #Client_context.full ->
+  t tzresult
+
 (** [shutdown t] properly shuts the layer 1 down. *)
 val shutdown : t -> unit Lwt.t